CAS-364: Refactor controller logic into helper functions
